docs: Production requirements
Motivated by our work with CDC for them to be able to internally host
nextstrain.org, but this documentation will also be useful for
ourselves.

We can adapt/amend it later to account for any "Nextstrain Groups
Server" (i.e. just the Groups bits of nextstrain.org) that we end up
creating.

Related-to: <https://github.com/nextstrain/private/issues/94>