config: Make COGNITO_USER_POOL_ID optional
When not given, the API endpoints which require it

    /groups/{name}/settings/members
    /groups/{name}/settings/roles/{role}/members
    /groups/{name}/settings/roles/{role}/members/{username}

return 503 Service Unavailable.

This is preferable to configuring a bogus value (to overcome the
previous value requirement) and the endpoints returning 500 Internal
Server Error.

We will likely still work on making Cognito pluggable here for other
user directories in the future.