Restore proxy from env support for AWS SDK v3 requests
The v2 SDK automatically used proxy settings in the environment (e.g.
http_proxy), but the v3 SDK does not.¹  There exists an aws-sdk-v3-proxy
package published by AWS², but it only supports HTTP/HTTPS proxies and
not SOCKS proxies.  SOCKS proxies are useful, so set up proxy-respecting
agents on the AWS clients ourselves.

Proxying is useful for development (e.g. to inspect requests), but also
required for some networks.  And there's an outstanding issue we'd like
to resolve which would require proxy support in production for us as
well.³  As proxies are a global, network-level concern, the environment
is the best place to configure them (and thus why the conventions
exist).

¹ <https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/node-configuring-proxies.html>
² <https://www.npmjs.com/package/aws-sdk-v3-proxy>
³ <https://github.com/nextstrain/private/issues/27>