authn: Store user "stale before" markers durably in SQLite when Redis is not available
Rather than non-durably in process memory.  This solves two issues:

 1. It allows for production use without Redis as long as certain other
    requirements are met.  I'll be documenting these and relaxing the
    Redis requirement in a subsequent commit.  This is motivated by
    CDC's deployment.¹

 2. It makes development more like production, which is important as the
    difference wasn't obvious and led to strange behaviour, e.g. across
    server restarts.²

I knew the transient in-memory store was not ideal when I wrote it, but
it was very expedient! and I didn't want to address it further at the
time.  This now rectifies that.

The configuration of the keyv library with a key namespace ensures that
no key name migration is needed with this change.

¹ <https://github.com/nextstrain/private/issues/94>
² <https://github.com/nextstrain/nextstrain.org/pull/581#discussion_r1351059469>