Allow and document how to run safely in production mode without Redis
This is made possible by the previous commit which switched the
non-Redis fallback from an in-memory store to an on-disk SQLite store.

As a nice side-effect, we can now set REDIS_REQUIRED=true in development
mode (i.e. not the default) to more easily test the production mode code
path.

Related-to: <https://github.com/nextstrain/private/issues/94>