Explicitly configure AWS region
Make it explicit in our config, code, and documentation.

This became newly and implicitly necessary with the upgrade from the v2
to v3 SDK in b326340b.  I didn't notice in development, however, because
region was implicitly sourced from the default profile in my personal
AWS config file.  That config file isn't available in production, of
course, and this resulted in some requests 500-ing with "region not set"
errors upon deploy to Heroku.¹  I likely could have noticed had I tried
the Heroku review app for the original changes.

Related-to: <https://github.com/nextstrain/nextstrain.org/pull/733>

¹ <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1698343822747429>