Add context to console error on groups page
Current implementation means an error for the settings page is shown in
the console for every visit to a group page where you aren't logged in
as an owner of the group.

It confused me until I realized that `canUserEditGroupSettings` is being
called on the group page to determine whether to display the edit
button. It seems like this is by design, and I'm not sure there is an
easy way to achieve the same behavior without sending this call to
/groups/{group}/settings. I think ok to leave as-is, but some extra
context might be helpful for anyone who has a tendency to think that
something is wrong when seeing red error messages in the browser
console.