[Github action] Rebuild resource collection index
See added documentation for corresponding AWS details. Given that the
resources all come from public-facing buckets (core & staging) it seems
ok to run this from a public repo, but we may want to revisit this once
we start consuming private data.

The index is only generated for datasets (not intermediate files) and
only for the core bucket (nextstrain-data) as the that's all that's
currently handled by the server, so it saves us a little s3 storage,
transfer overhead and server memory footprint. Future work
listing/visualising all available data will use this and so this
filtering is only temporary.