Use inline exports
This a follow-up to "Use cjs-to-es6 to help conversion" (1963fcd7),
which left exports at the bottom of the file. That was necessary for
CJS, but now ESM syntax can be leveraged for more concise code.

I searched for instances of "export {" and manually applied the
conversion where possible. I skipped src/async.js since that's a
vendored copy of an external script.