Add a "contact" page
Previously, the only means of contact shown on this website was the
hello@ email. This new contact page lists preferred means of
communication under various scenarios.

hello@ is still appropriate in the case of setting up a new group for
Nextstrain Groups, but all the other references to the email have been
replaced with a link to the new page.