Add script to update GitHub token on Heroku apps
Using a script allows for minimal time between GitHub token regeneration
and config var updates, especially when the same token is used across
multiple apps. Also useful to somewhat automate the process given GitHub
requires token regeneration at least once every year.

Note that it'll still take some time between manually regenerating the
token on the GitHub UI, pasting it as an argument to this script, then
running the script. In the case of this GitHub token, it's not currently
being used at a level where this matters critically.

In the future if this becomes more critical, the downtime could be
eliminated by generating a new token while keeping the old one valid.