aws/cognito: Extend validity of the CLI's refresh tokens from 30 → 90 days
In practice, 30 days seems too short.

I've observed users who get the CLI installed and logged in but then
don't get to uploading their datasets until more than 30 days later
(i.e. they spend time producing the datasets or on other work) and so
then have to log in again when it's time to upload.  This has been a
minor frustration and stumbling block for them.

In a different use case, @corneliusroemer has also found 30 days to be
too short in automation contexts.  Although I'd suggest a different
approach for automation¹, I suspect his approach will not be an uncommon
one for users to take so we should ensure it's not too onerous when they
do.

Resolves <https://github.com/nextstrain/cli/issues/337>.

¹ <https://github.com/nextstrain/cli/issues/337#issuecomment-1883534419>