[resource indexer] modify resource URLs
following the approach the server uses to redirect URLs. For instance,
we change the resource name for "dengue/denv1" to "dengue/denv1/genome"
because the former URL is redirected to the latter.

This allows the server to access historical datasets for the URL
"dengue/denv1/genome" which now include all "dengue/denv1"
("dengue_denv1.json") datasets.

Due to the way the server currently works this will not allow historical
version access for "dengue/denv1" URLs because the instantiation of the
Dataset (Resource) instance happens before the resolve/redirecting
happens, and the "dengue/denv1" dataset is considered invalid because
it's not present in the index _and_ a version descriptor is being
provided.