Refactor dataset path redirects & preserve queries
These are defined in a stand-alone function so that they can be easily
reused by the resource indexer to associate "old" filenames with their
new dataset path.

Express routes can include patterns and regexes but these are hard to
parse outside of express so I've converted to plain strings. It was
possible to parse the old routes via `path-to-regexp` or by
instantiating a small express app and iterating over the routing stack,
but both felt brittle and more complexity than its worth.

URL queries are now preserved for these redirects for all datasets. This
is a change in behaviour for the monkeypox URLs - as an example
"nextstrain.org/monkeypox/mpxv?c=region" will now redirect to
"nextstrain.org/mpox/all-clades?c=region". There is a very sizeable
caveat to this however, because if the redirect path described here is
subsequently redirected by our `canonicalizeDataset` middleware the
query will be discarded at that point. In practice, all of the ncov
paths and 1/4 of the mpox paths described here are subsequently
redirected and so the URL query is not preserved.
<https://github.com/nextstrain/nextstrain.org/issues/792> tracks this.