[resource indexer] use hardcoded redirects
Recent work surfaced previous dataset versions across URL redirects¹ by
mirroring the dataset-name resolution process we use for requests on the
server. However it neglected to consider the redirects which are handled
prior to this in the server. This commit adds that functionality as
well. This situation was recently discussed in slack².

To use a couple of examples representative of the redirects we use:
1. the dataset (URL) path "mpox/all-clades" now includes previous
   versions which were named "monkeypox_mpxv.json", thus extending the
   snapshot history of this dataset from 2023-09-23 to 2022-06-12.
2. the dataset (URL) path "ncov/gisaid/global/6m" (and paths which would
   be resolved to this, such as "ncov", "ncov/gisaid") contained periods
   without any snapshots because we were using the url "ncov/global".
   These datasets (versions of "ncov_global.json") are now included as
   snapshots for "ncov/gisaid/global/6m".

¹ <https://github.com/nextstrain/nextstrain.org/pull/783>
² <https://bedfordlab.slack.com/archives/CSKMU6YUC/p1706483980082939>