dev: Wait up to 60s for the Heroku build status to change to "succeeded"
It used to be that once the Heroku build log stream response completed,
their API would return a build status of "succeeded".  Recently however
it seems like now there's a delay between the log stream response
completing and their API returning the up-to-date build status.  Build
in a 60s buffer wait time to account for that.

Resolves <https://github.com/nextstrain/nextstrain.org/issues/794>.