[resource listing UI] implementation
This implements the basic UI/UX of resource display, including resource
grouping, filtering, sorting, and a pseudo-tree like display. Styles are
primarily chosen for laptop sized screens and above, but the page will
also display nicely on mobile.

A number of expediencies are taken, such as hardcoding the resource type
as 'dataset'. Future commits will add on functionality such as display
of snapshots.

The grouping of resources is two fold:
- (1) resources are grouped into pathogens, based on the first name in
  the resource path. For core datasets this works well, with the
  exception of 'flu' for which we have plans to split into
  'seasonal-flu' and 'avian-flu' at the dataset name level.
- (2) resources within each group are sorted by last updated date, with
  alphabetical sorting used for those with identical dates.

Resource groups display a limited number of individual resources, but
the user can expand to show all. The number displayed varies depending
on how many groups are displayed.