[resource listing UI] improve group info header
The main aim of these additions (quicklinks, making the group name a
link to the default dataset) is to surface the important datasets
quickly, especially for those pathogens which have dozens of datasets
(flu, ncov).

Similarly, allowing the group name to be customised helps
those who may not be familiar with our terminology (e.g. ncov is now not
used frequently).

Finally, we expose the SARS-CoV-2 pathogen page as a quick-link (chosen
after PR feedback). We have these for two pages (SC2 + flu) however the
flu page is simply tiles to a fixed set of datasets and is rendered
obsolete by this work on resource listing PR (although the page is not
modified here). If we expand these pages in the future we can revisit
the UI we use to surface them. this listing UI.