[resource listing UI] summarise update frequency
Uses a heuristic to summarise the frequency and regularity with which
resources have been updated. A previous prototype of this explored using
KDEs and sparklines to convey frequency however the UI became a little
too busy; this information is still available (in a better form) via the
modal to select snapshots.