Allow using built next.js assets
Running the nextstrain.org server in production mode is complicated by
the need for many environment variables. A common aim is to test/run the
front-end site (next.js) using pre-built / static-site-generated assets.
We expose the env variable `USE_PREBUILT_STATIC_SITE` to do this.

A (large!) side effect of this is that Next.JS will set the NODE_ENV to
'production' and so any code which uses this (as opposed to our own
`config.PRODUCTION` flag) will be confusing! One such case is the
`sslRedirect` middleware¹ which we now conditionally add based on
`config.PRODUCTION`.

Server start-up message inspired in part by
<https://github.com/nextstrain/nextstrain.org/issues/685>

<https://github.com/nextstrain/nextstrain.org/pull/810#discussion_r1556212860>
has discussion regarding the environment variable name.

¹ <https://github.com/paulomcnally/node-heroku-ssl-redirect/blob/87dc254783603e9801428fd2d5fa424136932971/src/index.ts#L9>