Remove unused/deprecated code
The search pages were not moved over to next.js in the previous commit
and have not worked for at least a year -- see the error banner added in
<https://github.com/nextstrain/nextstrain.org/commit/86956ef380ad0134d1fe18e40fdd18dadcaafa42>
for more context.

The Twitter-related JSX + CSS was not used in the code but is a relic
from a previous splash page.

The prisma CSS is also left over from a time where we hosted
documentation in the nextstrain.org static site (i.e. it predated
docs.nextstrain.org). This CSS (and the fira-mono font) was not moved
over to next.js.