Update static-site linting
These changes are in preparation for the subsequent commit which will
migrate to using Next.js for the static site.

As the static-site will no longer be its own siloed node app,
differentiating between `lint:server` and `lint:static-site` commands is
clearer, but the `lint` rule remains as an alias for `lint:server`.

For unknown reasons, the invocations 'eslint static-site' and 'cd
static-site && eslint .' behaved differently when next.js rules were
involved, with the former reporting a subset of the errors/warnings
picked up by the latter. The next.js config uses a lot of path
manipulation¹ so I presume that's the root cause, but the simple
solution is to change directory before invoking the rule.

¹ e.g. <https://github.com/vercel/next.js/blob/canary/packages/eslint-config-next/index.js>