Reinstate (and change) sars-cov-2 map
Switches to `react-map-gl` rather than the `react-mapbox-gl` library
which is not compatible with React 18 and hasn't been updated for 4
years. Most changes were straightforward, however the "clusters"
functionality¹ would need to be added by us. Comparing the output I
actually like the non-cluster UI better, although this is subjective,
and I don't feel it's necessary to add the code to reinstate the
clusters UI. While the current (and previous) map does a good job of
showing the geographic spread of SC2 datasets most have now fallen out
of date as genomic-epi has become less important, and the map (nor the
page in general) doesn't convey the update-date well. We should revisit
this.

¹ Previous API <https://github.com/alex3165/react-mapbox-gl/blob/master/docs/API.md#cluster>
and a similar UI implemented in react-map-gl:
<https://visgl.github.io/react-map-gl/examples/clusters>