Remove unused font weight 100
We defined font-weight:100 however it was being overridden by a child
element, so just remove it. (Actually rendering the text using 100 is
too thin.) Since we are never actually using weight 100 we no longer
fetch it. As part of this work I dug into the way we're using fonts and
compared the site as of this commit and the previous Gatsby site:

Gatsby made an external API request¹  for lato weights
{100,200,300,400,500,600,700} but of these only {100,300,400,700} are
actually valid². Nextstrain.org styles define
{100,200,300,400,500,800,900} font-weights, and when a specified weight
doesn't exist, a nearby weight is used³. In effect, this meant that the
following was happening:

100         (defined but never used)
200, 300 -> rendered using 300 (lato-light)
400, 500 -> rendered using 400 (lato-regular)
800, 900 -> rendered using 700 (lato-bold)

In the current next.js site, we are now requesting {300, 400, 700} for
both regular + italic. This should result in no changes except for
italic text which now uses the italic fontface rather than faux italic.
The real italic font appears a little more condensed.

This commit was prompted by PR comment / conversation
<https://github.com/nextstrain/nextstrain.org/pull/810#discussion_r1550626963>

¹ <https://fonts.googleapis.com/css?family=Lato: 100,200,300,400,500,600,700|Fira Mono: 400>

² <https://fonts.google.com/specimen/Lato>

³ <https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-weight>