Update tests re: missing resource pages
The previous tests relied on an implementation quirk of the Gatsby site.
When a resource page (e.g. "/flu/bad") didn't exist we (via express)
would explicitly send the Gatsby 404 page. Gatsby would then perform
client-side-routing to show the flu missing resource page (with an error
banner). Our tests were asserting the response matched the Gatsby 404
page's HTML which is technically true (they passed the tests!) but
that's not the page the user sees.

The current implementation doesn't have this quirk; in the above example
we directly send the appropriate flu page. One (very slight) downside is
that the response code for these pages is 200.

One difference between the previous Gatsby implementation and now is
that the initial next.js response is devoid of any useful HTML, with
subsequent requests adding the HTML which the user sees. There are a few
ways we can change this, but for testing purposes using something like
puppeteer to (e.g.) explicitly test for the error banner is probably a
better approach.