Pin NodeJS to 20.9.0
Consistent CI failures of `cryptography.test.js` were observed when
running on higher versions of NodeJS, including 20.11.1 and 20.12.0 (via
GitHub actions). As of 2024-04-07 NodeJS 20.9.0 is the latest available
in Conda (and is available across all architectures commonly used by
us), which is one reason I never saw this failure in development.
Pinning to 20.9.0 (as done in this commit) isn't a long term solution
and I'll create an issue to track this bug.

We also change the pinning mechanism for `actions/setup-node@v4` to use
the NodeJS version defined in `package.json`¹ which is how Heroku (via
heroku/nodejs buildpack)² resolves the nodejs version.

The PR thread where this was originally debugged / discussed:
<https://github.com/nextstrain/nextstrain.org/pull/810#discussion_r1553164104>

¹ <https://github.com/actions/setup-node/blob/main/docs/advanced-usage.md#node-version-file>
² <https://devcenter.heroku.com/articles/nodejs-support#specifying-a-node-js-version>