Enable plausible analytics
I investigated the plausible-recommended plugin¹ but for our usage we
only need a single <script> tag across the site. Tested on localhost by
using `script.local.js`²

We should consider proxying the script though our site as a proportion
(perhaps as high as 60%) of people will block this script, but that's
left as future work.

¹ <https://github.com/4lejandrito/next-plausible>
² <https://plausible.io/docs/script-extensions>
³ <https://plausible.io/docs/proxy/introduction>