WIP ssr for /team page
The page will be generated using Next.js SSR (only compiled once, at
build time, not when the server is running despite the first _S_ in SSR
standing for Server. At least, that's my understanding, but I need to
confirm this.)

Notice that _without_ JS (i.e. the HTML that's delivered to the browser)
the layout is completely wrong. With JS enabled you'll get a flash as
the browser hydrates and client side rendering updates the DOM to use
the correct layout. The layout difference is due to the styled
components styles not being included in CSS.