index-resources.yml: Create separate indexes for Heroku apps
We are running into out-of-sync issues between the resources index
and the production site when canary is not immediately promoted.

This change creates separate indexes for the nextstrain-server and
nextstrain-canary Heroku apps by checking out the commit of their latest
released slug.

The separate resources JSONs are uploaded to separate paths prefixed by
the app name, which required updating the permissions for
`GitHubActionsRoleResourceIndexer` outside of this commit.