Update core cards (main splash page)
To use the full paths rather than paths which will be redirected. I
changed to full paths to avoid what I felt was an unnecessary redirect
cycle. The upside is faster loading pages, although the actual
improvement may be tiny. The downside is decoupling the manifest
seasonal-flu default from the splash page tile dataset. (Although I see
this as another upside, not a downside.)