dev: Unbundle `next build static-site` from `npm run test:ci` and `… dev:ssg`
As someone who wants to routinely use the actual built static-site¹, it
makes no sense to pay the time cost of `next build` every run of the
test suite or restart of the dev server.  We already check that the
built assets are available at server start, so if you forget you do get
a reminder (in the form of an error).

¹ The hot-reloading static-site is great for working on the static-site
  itself, but when working on the server-side routes, I'd rather have
  higher fidelity with production.