Start using versioned `RESOURCE_INDEX`
We've run into out-of-sync issues between the resources index and the
production site when canary is not immediately promoted. This change
starts using versioned indexes so that changes to the canary resource
index will not affect the production site.

This required updating the permissions for `GitHubActionsRoleResourceIndexer`
outside of this commit.