fetch: Log with debug() instead of our utils.verbose()
This module is now fully independent of the rest of our codebase,
relying only on third-party deps.  That makes it suitable for use in
very early process start up, before all modules and globals are
initialized (e.g. global.verbose).

I'm continuing to use the "debug" module as I opted to in "authn: Add
debug logging for user restoration from sessions" (3cf94a7e), for
largely the same reasons and with the same caveats.

A result of this change is that --verbose no longer includes fetch
logging.  I adjust the invocations for 1) the Heroku process and 2) the
`npm run test:ci` alias to preserve logging of fetches in those places
and avoid the loss of a useful behaviour.  In development and other
contexts, a similar DEBUG=nextstrain:fetch value will be necessary if
fetch logging is still desired.  It's pretty chatty so is now left up to
the developer's preferences.