config: Use our own fetch() for logging and proxy (and caching) support
The logging and proxy support is particularly useful for debugging and
troubleshooting.  Also, the previous lack of proxy support here
complicated running the server under a proxy for inspecting other
requests, as the OIDC configuration endpoint had to be excluded else
it'd hit certificate errors.

The caching support is not used here at the moment because the only
thing we fetch here, the OIDC configuration, is not cacheable.

Motivated by wanting to observe the AWS STS GetCallerIdentity request
now made on server start up¹ and realizing that what used to work
(configuring a proxy) did not due to the aforementioned complication.

¹ "[startup message] show IAM user name" (bddf2254)