dev: Allow manually-invoked deploys to a named Heroku app
This is handy for testing the "deploy" job and potentially for deploying
a larger work-in-progress change to dev.nextstrain.org (i.e. before its
ready for next.nextstrain.org).

The "deploy" job now references a GitHub environment¹ named after the
Heroku app to access Heroku API credentials and record the deployment.
I've provisioned the "nextstrain-canary"² and "nextstrain-dev"³
environments in the repo settings.  Each has the same Heroku API token
attached to bot@nextstrain.org, scoped down to "write-protected"
(instead of the default of "global").⁴  The "nextstrain-canary"
environment can only be used when running on master and so is only
usable by those who can push to master.  The "nextstrain-dev"
environment, however, can be used from any branch, and thus as a
safeguard requires manual approval of the deployment by
@nextstrain/core.  Unfortunately the Heroku API tokens can't be scoped
down to a single app instead; they always allow access to all apps the
bot@ account can manage.

¹ <https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment>
² <https://github.com/nextstrain/nextstrain.org/settings/environments/440351004/edit>
³ <https://github.com/nextstrain/nextstrain.org/settings/environments/13941227/edit>
⁴ <https://devcenter.heroku.com/articles/oauth#scopes>