dev: Work around the ancient Bash 3.2.57 on macOS
@victorlin noted in review that the previous construct threw an "unbound
variable" error on macOS's system Bash 3.2.57.¹  This is because after
3.2.57 and at least by 4.4.20 (but maybe earlier), Bash stopped treating
the "@" and "*" parameters as unset under -u when empty.²

Use an array slice³ (i.e. substring expansion, ${parameter:offset},
applied to an array) as a workaround.  Note that testing if
"docker_args" is empty, e.g. "${docker_args:+${docker_args[@]}}", won't
work because it leaves us with an empty string argument when unset.

¹ <https://github.com/nextstrain/nextstrain.org/pull/857#discussion_r1597296074>
² <https://github.com/nextstrain/ingest/pull/17#discussion_r1312013780>
³ <https://www.gnu.org/software/bash/manual/bash.html#Shell-Parameter-Expansion>