index-resources: update `IS_WORKFLOW_CALL` to default to false
When the workflow is not triggered by `workflow_call`, then the
`is-workflow-call` input sets the `IS_WORKFLOW_CALL` environment
variable to an empty string. This commit explicitly sets the
default value for `IS_WORKFLOW_CALL` to `false` so that
`./scripts/get-resource-index-ref-matrix` can run the appropriate steps.

Follow up to https://github.com/nextstrain/nextstrain.org/pull/865.