Left-align the cards
This is the default behavior of justify-content so an alternative is to
simply remove the line, but better to be explicit.

Note: there is a gap on the right side of the cards because the
collection of cards can't be centered with this approach. There are
workarounds¹ that don't use flexbox, but keeping it like this for now.

¹ <https://stackoverflow.com/q/32802202>