Create custom InternalLink component
Currently, internal links are represented by a mix of Next.js Link and
HTML anchor. It should be either one or the other across the board.

Create a component to standardize this and allow for easy switching
between the two depending on if we want client side rendering or server
side rendering.

Start by making this a Next.js Link since that's more commonly used.