Remove `versioned` attribute from Resource interface
This attribute only served as an alias to determine if the optional
attributes `firstUpdated|dates|nVersions|updateCadence` are available.
It's more direct (and preferred by the compiler) to check for presence
of the individual attributes directly where needed.

`versioned` at the top level on ListResourcesResponsiveProps is still
valuable and should be kept.