Swap splash tiles to viz screenshots
This commit swaps the tiles used on the splash page from the previous doctored electron micrographs to be simple screenshots of the pages / pathogens that are linked to. This has two advantages:
1. It's super simple and no longer requires sourcing a copyright-free image and doctoring it in LunaPic
2. It's a more up front, more transparent window into what you get when you click on a tile

I used tree vs map depending on spatial / epi patterns. For pathogens like WNV and Ebola, the map is meaningful and interesting, while for things like H3N2, it's less relevant. I generally tried to stick close to the content that's directly linked and not update coloring, etc..., but I made a couple of exceptions for visual interest.