Use absolute value for gap
The percentage value isn't factored in to clientHeight, which is a value
that is used to calculate the total height of the expanded showcase.

This subtlety isn't very noticeable with 1% + white space on the border
of the current CardComponent. It is more apparent at larger percents and
with other work involving a CardComponent that uses a colored border,
which was being cut off in the expanded showcase.