Swap core pathogens tiles to viz screenshots
The swap makes less sense here since clicking does not immediately go to
the visualization.

However, this is the last place that the electron micrographs are used.
Removing them from here removes the need to maintain two separate images
per pathogen and removes the need to conditionally include image credit
in the footer.

Screenshots added for pathogens that weren't included in splash.