Move concurrency limit to the workflow level
The concurrency limit on the deploy step was meant to prevent multiple
deployment processes from running at the same time. However, most of the
time spent on the deployment process is in the build step, which means
there is not likely to be any overlap on the deploy step.

Set the limit on the workflow level to capture the entire deployment
process. This will also affect runs that don't deploy (e.g. PR updates),
but most of the time only the latest push to a PR is important so it's
fine that older runs are canceled.