Remove unused `location` props
I was looking into tackling the TODO comment for the location prop¹ in
the groups JS routing. The location prop was being passed around to be
used in the NavBar component, but the use of it was removed in a
recent refactoring.² So just remove all of the unused `location` props
for the `NavBar` and the `GenericPage` components.

¹ <https://github.com/nextstrain/nextstrain.org/blob/1aeb1e499b9f25d9a918465fd3450fca51f40a2d/static-site/pages/groups/%5B...groupName%5D.jsx#L22>
² <https://github.com/nextstrain/nextstrain.org/commit/ac12e3201a95c7e665ac6165f4f6b1f206207eb4>