Add page for listing Group members
The new page is available at `groups/<groupName>/settings/members`.
I chose to use the same URL as the API so that we don't have to take
up another namespace (e.g. `groups/<groupName>/members`). The main Group
page will only link out to the members page if the current user is
authorized to list Group members.

The members list only displays the most privileged role for each member
to keep the display simple. It assumes the group roles API returns the
roles in the order from least to most privileged.