[provision group] Check changes against cognito
The provision-users command requires us to provide both a username and
email for each user we want to add. Commonly we don't know the username
for a given email, so one usage of this script is designed around
listing existing users (and their roles) for a given email. A second
use of the script is cross-referencing the proposed username/email/role
against the existing pool.

Note that emails & usernames in our pools are _case sensitive_
<https://github.com/nextstrain/nextstrain.org/issues/1258> which adds
a lot of complexity to this script, as one aim is to avoid us
provisioning new users which differ from an existing user only by case.

The output is perhaps overly verbose, but given the frequency we add
users I find this verbosity clarifying rather than annoying.